Changed app font Added fragment for teacher, classes. Added dialog window for specific participant in ClassFragment Theme can only be light(for easier design) Working version

+++ app/src/main/java/com/policarp/journal/database/response/entities/StudentEntity.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class StudentEntity {
-    private Integer studentId;
-    private Integer participantId;
-    private Integer attachedClassId;
+    private Long studentId;
+    private Long participantId;
+    private Long attachedClassId;
     @Override
     public String toString() {
         return "Student{" +
